requirements.txt
flask
flask-cors
python-dotenv
openai
langchain
chromadb
beautifulsoup4
requests
PyPDF2
tiktoken
sentence-transformers
langchain-community
pypdf